+++ src/maal/income.rs
@@ -145,6 +145,10 @@ impl IncomeZakatCalculator {
 }
 
 impl CalculateZakat for IncomeZakatCalculator {
+    fn validate_input(&self) -> Result<(), ZakatError> {
+        self.validate()
+    }
+
     fn calculate_zakat<C: ZakatConfigArgument>(&self, config: C) -> Result<ZakatDetails, ZakatError> {
         self.validate()?;
         let config_cow = config.resolve_config();